% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFPWM.R
\name{writeFPWM}
\alias{writeFPWM}
\title{Writes the FPWM in transfac or FPWMtransfac format}
\usage{
writeFPWM(FPWM = NULL, format = "transfac", fileName = NULL)
}
\arguments{
\item{FPWM}{FPWM object.}

\item{format}{[character] the "transfac" option will output a standard transfact matrix per binding partner. The "FPWMtransfac" prints a single matrix with all the binding partners in FPWM format.}

\item{fileName}{[character] name of the file where the FPWM is going to be written.}
}
\value{
A plain text file with the FPWM in transfac or FPWMtransfac format.
}
\description{
This function saves the FPWM in transfac format.
}
\examples{
fpwm <- createFPWM(mainTF ="CEBPB",partners = c("ATF4","ATF7","ATF3"), cell = "K562", forkPosition = 5)
writeFPWM(FPWM = fpwm, format = "transfac", fileName = "FPWM.transfact" )
writeFPWM(FPWM = fpwm, format = "FPWMtransfac", fileName = "FPWM.FPWMtransfac" )
}
\keyword{writeFPWM}
